# crates/api/src/error.rs
//! Consistent error envelope (design §6.1): every non-2xx response is
//! `{"error": {"code", "message"}}` — the same shape on 400, 404 and 500.

use axum::Json;
use axum::http::StatusCode;
use axum::response::{IntoResponse, Response};
use serde::Serialize;
use utoipa::ToSchema;

/// The error envelope every non-2xx response carries.
#[derive(Debug, Serialize, ToSchema)]
pub struct ErrorBody {
    /// The error itself.
    pub error: ErrorDetail,
}

/// Machine-readable error code plus a human-readable message.
#[derive(Debug, Serialize, ToSchema)]
pub struct ErrorDetail {
    /// Stable machine-readable code (e.g. `invalid_cursor`, `not_found`).
    #[schema(examples("invalid_cursor"))]
    pub code: String,
    /// Human-readable explanation.
    pub message: String,
}

/// Handler-level API failure; renders as the [`ErrorBody`] envelope.
#[derive(Debug)]
pub enum ApiError {
    /// Malformed client input (bad cursor, bad limit, bad query value).
    BadRequest {
        /// Stable machine-readable code.
        code: &'static str,
        /// Human-readable explanation.
        message: String,
    },
    /// The addressed resource does not exist.
    NotFound {
        /// Human-readable explanation.
        message: String,
    },
    /// The request is well-formed but conflicts with current state (e.g.
    /// resolving an already-resolved review task).
    Conflict {
        /// Stable machine-readable code.
        code: &'static str,
        /// Human-readable explanation.
        message: String,
    },
    /// Missing or invalid credentials (bad/revoked key, bad admin token).
    Unauthorized {
        /// Stable machine-readable code.
        code: &'static str,
        /// Human-readable explanation.
        message: String,
    },
    /// Authenticated, but the tier does not include this capability.
    Forbidden {
        /// Stable machine-readable code.
        code: &'static str,
        /// Human-readable explanation.
        message: String,
    },
    /// Quota or rate limit exhausted (goal 050 metering).
    TooManyRequests {
        /// Stable machine-readable code.
        code: &'static str,
        /// Human-readable explanation.
        message: String,
    },
    /// A required backend integration is not configured (fail closed).
    Unavailable {
        /// Stable machine-readable code.
        code: &'static str,
        /// Human-readable explanation.
        message: String,
    },
    /// Anything the client cannot fix; details stay server-side.
    Internal(anyhow::Error),
}

impl ApiError {
    /// Shorthand for a [`ApiError::BadRequest`].
    pub fn bad_request(code: &'static str, message: impl Into<String>) -> Self {
        Self::BadRequest {
            code,
            message: message.into(),
        }
    }

    /// Shorthand for a [`ApiError::Unauthorized`].
    pub fn unauthorized(code: &'static str, message: impl Into<String>) -> Self {
        Self::Unauthorized {
            code,
            message: message.into(),
        }
    }

    /// Shorthand for a [`ApiError::Forbidden`].
    pub fn forbidden(code: &'static str, message: impl Into<String>) -> Self {
        Self::Forbidden {
            code,
            message: message.into(),
        }
    }

    /// Shorthand for a [`ApiError::TooManyRequests`].
    pub fn too_many_requests(code: &'static str, message: impl Into<String>) -> Self {
        Self::TooManyRequests {
            code,
            message: message.into(),
        }
    }
}

impl From<anyhow::Error> for ApiError {
    fn from(err: anyhow::Error) -> Self {
        Self::Internal(err)
    }
}

impl From<sqlx::Error> for ApiError {
    fn from(err: sqlx::Error) -> Self {
        Self::Internal(err.into())
    }
}

impl From<govfolio_core::query::QueryError> for ApiError {
    fn from(err: govfolio_core::query::QueryError) -> Self {
        // Grammar values arrive pre-validated by serde; a token failure is
        // internal misuse, never client input.
        Self::Internal(err.into())
    }
}

impl IntoResponse for ApiError {
    fn into_response(self) -> Response {
        let (status, code, message) = match self {
            Self::BadRequest { code, message } => {
                (StatusCode::BAD_REQUEST, code.to_owned(), message)
            }
            Self::NotFound { message } => (StatusCode::NOT_FOUND, "not_found".to_owned(), message),
            Self::Conflict { code, message } => (StatusCode::CONFLICT, code.to_owned(), message),
            Self::Unauthorized { code, message } => {
                (StatusCode::UNAUTHORIZED, code.to_owned(), message)
            }
            Self::Forbidden { code, message } => (StatusCode::FORBIDDEN, code.to_owned(), message),
            Self::TooManyRequests { code, message } => {
                (StatusCode::TOO_MANY_REQUESTS, code.to_owned(), message)
            }
            Self::Unavailable { code, message } => {
                (StatusCode::SERVICE_UNAVAILABLE, code.to_owned(), message)
            }
            Self::Internal(err) => {
                // Details stay server-side; the envelope stays generic.
                eprintln!("internal error: {err:#}");
                (
                    StatusCode::INTERNAL_SERVER_ERROR,
                    "internal".to_owned(),
                    "internal server error".to_owned(),
                )
            }
        };
        let body = ErrorBody {
            error: ErrorDetail { code, message },
        };
        (status, Json(body)).into_response()
    }
}