# packages/contracts/openapi.json
{
  "components": {
    "schemas": {
...
      "ApiKeyInfo": {
        "description": "One key as listed — metadata only, nothing secret-derived.",
        "properties": {
          "created_at": {
            "description": "When the key was created.",
            "format": "date-time",
            "type": "string"
          },
          "id": {
            "description": "Key ULID.",
            "pattern": "^[0-7][0-9A-HJKMNP-TV-Z]{25}$",
            "type": "string"
          },
          "label": {
            "description": "Human label.",
            "type": "string"
          },
          "revoked_at": {
            "description": "When the key was revoked; `null` while active.",
            "format": "date-time",
            "type": [
              "string",
              "null"
            ]
          }
        },
        "required": [
          "id",
          "label",
          "created_at"
        ],
        "type": "object"
      },
...
      "CreateKey": {
        "description": "Body of `POST /v1/keys`.",
        "properties": {
          "label": {
            "description": "Human label for the key (shown in listings).",
            "type": "string"
          },
          "user_id": {
            "description": "Target user — REQUIRED with admin auth, FORBIDDEN with key auth\n(a key only mints keys for its own account).",
            "type": [
              "string",
              "null"
            ]
          }
        },
        "required": [
          "label"
        ],
        "type": "object"
      },
      "CreateUser": {
        "description": "Body of `POST /v1/users` (admin bootstrap).",
        "properties": {
          "email": {
            "description": "Account email (unique).",
            "type": "string"
          },
          "tier": {
            "oneOf": [
              {
                "type": "null"
              },
              {
                "$ref": "#/components/schemas/Tier",
                "description": "Tier; defaults to `free`."
              }
            ]
          }
        },
        "required": [
          "email"
        ],
        "type": "object"
      },
      "CreatedKey": {
        "description": "The creation response — the ONLY place the plaintext key ever appears.",
        "properties": {
          "created_at": {
            "description": "When the key was created.",
            "format": "date-time",
            "type": "string"
          },
          "id": {
            "description": "Key ULID (use for revocation).",
            "pattern": "^[0-7][0-9A-HJKMNP-TV-Z]{25}$",
            "type": "string"
          },
          "key": {
            "description": "The bearer token (`gfk_...`). Shown once; store it now — only its\nhash is retained.",
            "type": "string"
          },
          "label": {
            "description": "Human label.",
            "type": "string"
          }
        },
        "required": [
          "id",
          "key",
          "label",
          "created_at"
        ],
        "type": "object"
      },
...
      "ErrorBody": {
        "description": "The error envelope every non-2xx response carries.",
        "properties": {
          "error": {
            "$ref": "#/components/schemas/ErrorDetail",
            "description": "The error itself."
          }
        },
        "required": [
          "error"
        ],
        "type": "object"
      },
      "ErrorDetail": {
        "description": "Machine-readable error code plus a human-readable message.",
        "properties": {
          "code": {
            "description": "Stable machine-readable code (e.g. `invalid_cursor`, `not_found`).",
            "examples": [
              "invalid_cursor"
            ],
            "type": "string"
          },
          "message": {
            "description": "Human-readable explanation.",
            "type": "string"
          }
        },
        "required": [
          "code",
          "message"
        ],
        "type": "object"
      },
...
      "Tier": {
        "description": "`user_account.tier` CHECK (design §6.2 freemium table): the tier decides\nfreshness (free = 24h-delayed via the ONE visibility bound in\n`core::query`) and the daily request quota. Default is `free`.",
        "enum": [
          "free",
          "pro",
          "data"
        ],
        "type": "string"
      },
      "UserAccount": {
        "description": "One account (Gold-side `user_account`).",
        "properties": {
          "created_at": {
            "description": "When the account was created.",
            "format": "date-time",
            "type": "string"
          },
          "email": {
            "description": "Unique account email.",
            "type": "string"
          },
          "id": {
            "description": "User ULID.",
            "pattern": "^[0-7][0-9A-HJKMNP-TV-Z]{25}$",
            "type": "string"
          },
          "tier": {
            "$ref": "#/components/schemas/Tier",
            "description": "Tier (decides freshness delay + daily quota, design §6.2)."
          }
        },
        "required": [
          "id",
          "email",
          "tier",
          "created_at"
        ],
        "type": "object"
      },
...
    }
  },
  "info": {
    "description": "Worldwide politician financial-disclosure tracking. Cursor pagination on ULIDs; consistent error envelope; verification_state on every record. Every successful GET carries a strong ETag (sha256 of the body); requests with a matching If-None-Match receive 304 Not Modified. Authentication: `Authorization: Bearer gfk_...` API keys (create via /v1/keys; hashes only at rest). Requests without a key are served at the free tier. Tiers control record freshness (free: filings we discovered less than 24 hours ago are not yet visible; pro/data: real time) and the daily request quota (counted per account; over quota is 429 quota_exceeded in the standard envelope). Anonymous traffic has a light per-IP rate backstop (429 rate_limited).",
    "title": "govfolio API",
    "version": "0.1.0"
  },
  "openapi": "3.1.0",
  "paths": {
...
    "/v1/keys": {
      "get": {
        "description": "# Errors\n`401` without a valid API key; `500` on backend failure.",
        "operationId": "list_keys",
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "$ref": "#/components/schemas/ApiKeyInfo"
                  },
                  "type": "array"
                }
              }
            },
            "description": "The caller's keys, metadata only"
          },
          "401": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorBody"
                }
              }
            },
            "description": "No valid API key"
          },
          "500": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorBody"
                }
              }
            },
            "description": "Internal error"
          }
        },
        "summary": "Lists the caller's keys (metadata only — the plaintext is unrecoverable\nby design).",
        "tags": [
          "account"
        ]
      },
      "post": {
        "description": "# Errors\n`401` without any valid credential; `400` on a blank label or a body/auth\nmismatch; `404` for an unknown user (admin path); `500` on backend\nfailure.",
        "operationId": "create_key",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateKey"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CreatedKey"
                }
              }
            },
            "description": "The created key, plaintext INCLUDED — shown exactly once"
          },
          "400": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorBody"
                }
              }
            },
            "description": "Blank label or user_id/auth mismatch"
          },
          "401": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorBody"
                }
              }
            },
            "description": "No valid API key or admin token"
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorBody"
                }
              }
            },
            "description": "Unknown user (admin path)"
          },
          "500": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorBody"
                }
              }
            },
            "description": "Internal error"
          }
        },
        "summary": "Creates an API key. With key auth: for the caller's own account. With\n`X-Admin-Token`: for the user named in the body (bootstrap).",
        "tags": [
          "account"
        ]
      }
    },
    "/v1/keys/{id}": {
      "delete": {
        "description": "# Errors\n`401` without any valid credential; `404` for an unknown (or not owned)\nkey; `500` on backend failure.",
        "operationId": "revoke_key",
        "parameters": [
          {
            "description": "Key ULID",
            "in": "path",
            "name": "id",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "Revoked (idempotent)"
          },
          "401": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorBody"
                }
              }
            },
            "description": "No valid API key or admin token"
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorBody"
                }
              }
            },
            "description": "Unknown or not owned key"
          },
          "500": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorBody"
                }
              }
            },
            "description": "Internal error"
          }
        },
        "summary": "Revokes a key (immediate — the very next request with it is `401`).\nKey auth revokes own keys; `X-Admin-Token` revokes any.",
        "tags": [
          "account"
        ]
      }
    },
...
    "/v1/users": {
      "post": {
        "description": "# Errors\n`401` without a valid `X-Admin-Token`; `400` on a blank email; `409` on a\nduplicate email; `500` on backend failure.",
        "operationId": "create_user",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateUser"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/UserAccount"
                }
              }
            },
            "description": "The created account"
          },
          "400": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorBody"
                }
              }
            },
            "description": "Blank email"
          },
          "401": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorBody"
                }
              }
            },
            "description": "Missing or invalid X-Admin-Token"
          },
          "409": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorBody"
                }
              }
            },
            "description": "Email already registered"
          },
          "500": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorBody"
                }
              }
            },
            "description": "Internal error"
          }
        },
        "summary": "Creates an account (bootstrap; real signup is deferred — see module docs).",
        "tags": [
          "account"
        ]
      }
    }
  },
...
}